backend/requirements.txt
cffi
click
colorama
cryptography
Flask
Flask-Cors
Flask-SQLAlchemy
greenlet
itsdangerous
Jinja2
MarkupSafe
pycparser
PyJWT
python-dotenv
python-http-client
sendgrid
six
SQLAlchemy
starkbank-ecdsa
Werkzeug